requirement.txt
langchain_core
python-dotenv
streamlit 
langchain_community
chromadb
faiss-cpu
langchain-ollama
langchainhub
groq
langchain-groq
beautifulsoup4

